Seperated emulator router and rag router
